requirements.txt
streamlit
pillow
numpy
opencv-python
rembg
onnxruntime
imageio
